Move top-level APIs to dry/logger + YARD
